ikev2: in ikev2_replace() capture Child's policy when replacing IKE
It was being set to LEMPTY (older code hat it set to .st_policy &
~POLICY_IPSEC_MASK but that can be reduced to LEMPTY).  The IKE SA
initiate code would then interpret LEMPTY as a no-ipsec-policy and
skip scheduling itself as the first child.

Fortunately the connection was also on the revival queue and when that
kicked in it would schedule the connection anyway.

Note that, due to quirks in pending.[hc], this changes the order
that the reviving connections are created.  Per:
   give pending a proper linked list #1573
an initiating IKE SA should probably give priority to its own
connection.